[NXH2004-17760] feat(clock_config): Allow override of external clock frequency
A board with an LPC might have an external clock generator which is
supplying a clock to the LPC.
Most LPCXpresso boards have a 16MHz XTAL fitted and this clock is
connected to the LPC.

A customer or other NXP board might have a XTAL with a different
frequency connected to it.
F.e. the [NXH3675EVK board](https://www.nxp.com/products/wireless-connectivity/bluetooth-le-audio/nxh3675evk-evaluation-kit-bluetooth-5-3-aoble-certified-with-lpc55-mcu-for-usb-audio:NXH3675EVK)
has an LPC5516 mounted with an XTAL of 32MHz.

At board-initialization, SystemCoreClockUpdate() is called to update
the core-clock with the relevant settings of the board.
Currently, when an external clock configuration is selected, we will
take the hardcoded CLK_CLK_IN value as the system core clock.

If a board has a different XTAL frequency than the hardcoded
CLK_CLK_IN, the wrong clock frequency is set resulting in several
issues (no booting, wrong uart prints, ...).

This results in the initial board bringup to require changes to the
core LPC files which should be SOC-specific and board-agnostic. Or it
would require the customer/user to wrap SystemCoreClockUpdate() in a
custom function selecting the correct clock frequency, which is rather cumbersome.

This commit adds an override feature to the external clock frequency
declaration.
- If CLK_CLK_IN is not specified in the project or command-line,
  no changes are done to the original behaviour.
- If CLK_CLK_IN is specified in the project or command-line,
  the user-specified frequency is selected.

This eases initial board-bringup of new LPC-based boards and allows easier integration with the NXP SDK.